Fixed the socket re-connecting problem for now
