+++ src/pages/community/[communityId].js
@@ -1,22 +1,24 @@
 import { getCommunityInfo } from "@/lib/api-calls/community";
 import { fetchMessages } from "@/lib/api-calls/messages";
+import { socket } from "@/lib/socket-client";
 import CommunityTopBar from "@/src/components/community/TopBar";
 import ScrollableMessageBox from "@/src/components/messages";
 import { Progress, useToast } from "@chakra-ui/react";
-import { useQueries } from "@tanstack/react-query";
+import { useQueries, useQueryClient } from "@tanstack/react-query";
 import { useSession } from "next-auth/react";
 import dynamic from "next/dynamic";
 import Head from "next/head";
 import { useRouter } from "next/router";
+import { useEffect } from "react";
 
 const Community = () => {
   const router = useRouter();
   const session = useSession();
   const toast = useToast();
+  const queryClient = useQueryClient();
+
   // Get the community id from the URL of the dynamic route in NextJS
   const { communityId } = router.query;
-
-  console.log("hello");
 
   const [
     { data: communityData, isLoading: isCommLoading, error },
@@ -39,6 +41,24 @@ const Community = () => {
       },
     ],
   });
+
+  useEffect(() => {
+    console.log("SOCKET CONNECTED");
+    socket.connect();
+    if (communityId) {
+      socket.on(`community-${communityId}`, (data) => {
+        queryClient.setQueryData(["messages", communityId], (prev) => [
+          ...prev,
+          data,
+        ]);
+      });
+    }
+    return () => {
+      socket.disconnect();
+      if (communityId) socket.off(`community-${communityId}`);
+      console.log("SOCKET DISCONNECTED");
+    };
+  }, [communityId]);
 
   if (error) {
     const { message, redirect } = error.response.data;
@@ -84,7 +104,7 @@ const Community = () => {
           <Progress size="md" isIndeterminate colorScheme="purple" />
         </div>
       ) : (
-        <div className="flex grow flex-col gap-3">
+        <div className="flex grow flex-col justify-between">
           <CommunityTopBar
             data={{
               ...communityData,
